Add note about upgrading with Composer
I use composer for new projects and wanted to also use it to download a fresh copy of Kirby 3 to update an old site. I used Composer to download in a directory next to the existing site, then followed steps 1, 2, and 3 in Replacing core files and folders. The one important step I missed was bringing over the `/vendor` folder, which made it all work again.